+++ pypika_tortoise/__init__.py
@@ -1,6 +1,6 @@
-from pypika.dialects import MSSQLQuery, MySQLQuery, OracleQuery, PostgreSQLQuery, SQLLiteQuery
-from pypika.enums import DatePart, Dialects, JoinType, Order
-from pypika.exceptions import (
+from .dialects import MSSQLQuery, MySQLQuery, OracleQuery, PostgreSQLQuery, SQLLiteQuery
+from .enums import DatePart, Dialects, JoinType, Order
+from .exceptions import (
     CaseException,
     FunctionException,
     GroupingException,
@@ -9,10 +9,10 @@ from pypika.exceptions import (
     RollupException,
     SetOperationException,
 )
-from pypika.queries import AliasedQuery, Column, Database, Query, Schema, Table
-from pypika.queries import make_columns as Columns
-from pypika.queries import make_tables as Tables
-from pypika.terms import (
+from .queries import AliasedQuery, Column, Database, Query, Schema, Table
+from .queries import make_columns as Columns
+from .queries import make_tables as Tables
+from .terms import (
     JSON,
     Array,
     Bracket,
