+++ pypika_tortoise/analytics.py
@@ -2,7 +2,7 @@
 Package for SQL analytic functions wrappers
 """
 
-from pypika.terms import AnalyticFunction, IgnoreNullsAnalyticFunction, WindowFrameAnalyticFunction
+from .terms import AnalyticFunction, IgnoreNullsAnalyticFunction, WindowFrameAnalyticFunction
 
 
 class Preceding(WindowFrameAnalyticFunction.Edge):
