+++ pypika_tortoise/dialects/oracle.py
@@ -1,7 +1,7 @@
 from typing import Any
 
-from pypika.enums import Dialects
-from pypika.queries import Query, QueryBuilder
+from ..enums import Dialects
+from ..queries import Query, QueryBuilder
 
 
 class OracleQuery(Query):
